requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.9.0=pypi_0
astor=0.8.1=pypi_0
blas=1.0=mkl
boto3=1.10.47=pypi_0
botocore=1.13.47=pypi_0
ca-certificates=2019.11.27=0
cachetools=4.0.0=pypi_0
certifi=2019.11.28=py37_0
cffi=1.13.2=py37h7a1dbc1_0
chardet=3.0.4=pypi_0
click=7.0=py37_0
cudatoolkit=10.1.243=h74a9793_0
docutils=0.15.2=pypi_0
dominate=2.4.0=pypi_0
flask=1.1.1=py_0
flask-bootstrap=3.3.7.1=pypi_0
flask-wtf=0.14.2=pypi_0
freetype=2.9.1=ha9979f8_1
future=0.18.2=pypi_0
gast=0.2.2=pypi_0
google-auth=1.10.0=pypi_0
google-auth-oauthlib=0.4.1=pypi_0
google-pasta=0.1.8=pypi_0
grpcio=1.26.0=pypi_0
h5py=2.10.0=pypi_0
icc_rt=2019.0.0=h0cc432a_1
idna=2.8=pypi_0
intel-openmp=2019.4=245
itsdangerous=1.1.0=py37_0
jinja2=2.10.3=py_0
jmespath=0.9.4=pypi_0
jpeg=9b=hb83a4c4_2
keras=2.3.1=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.0=pypi_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.1.0=h56a325e_0
markdown=3.1.1=pypi_0
markupsafe=1.1.1=py37he774522_0
mkl=2019.4=245
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.0.15=py37h14836fe_0
mkl_random=1.1.0=py37h675688f_0
ninja=1.9.0=py37h74a9793_0
numpy=1.17.4=py37h4320e6b_0
numpy-base=1.17.4=py37hc3f5095_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py37_0
openssl=1.1.1d=he774522_3
opt-einsum=3.1.0=pypi_0
pandas=0.25.3=py37ha925a31_0
pillow=6.2.1=py37hdc69c19_0
pip=19.3.1=py37_0
protobuf=3.11.2=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.7=pypi_0
pycparser=2.19=py37_0
python=3.7.5=h8c8aaf0_0
python-dateutil=2.8.1=py_0
python-twitter=3.5=pypi_0
pytorch=1.3.1=py3.7_cuda101_cudnn7_0
pytorch-pretrained-bert=0.6.2=pypi_0
pytz=2019.3=py_0
pyyaml=5.3=pypi_0
regex=2020.1.7=pypi_0
requests=2.22.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.0=pypi_0
s3transfer=0.2.1=pypi_0
scipy=1.4.1=pypi_0
setuptools=44.0.0=py37_0
six=1.13.0=py37_0
sqlite=3.30.1=he774522_0
tensorboard=2.0.2=pypi_0
tensorflow-estimator=2.0.1=pypi_0
tensorflow-gpu=2.0.0=pypi_0
termcolor=1.1.0=pypi_0
tk=8.6.8=hfa6e2cd_0
torchvision=0.4.2=py37_cu101
tqdm=4.41.1=pypi_0
twython=3.7.0=pypi_0
urllib3=1.25.7=pypi_0
vc=14.1=h0510ff6_4
visitor=0.1.3=pypi_0
vs2015_runtime=14.16.27012=hf0eaf9b_1
werkzeug=0.16.0=py_0
wheel=0.33.6=py37_0
wincertstore=0.2=py37_0
wrapt=1.11.2=pypi_0
wtforms=2.2.1=py37_0
xz=5.2.4=h2fa13f4_4
zlib=1.2.11=h62dcd97_3
zstd=1.3.7=h508b16e_0
